requirements.txt
boto3==1.28.38
requests==2.31.0
python-dotenv==1.0.0
line-bot-sdk==3.1.0
pytz==2023.3
google-generativeai==0.3.1
